atividade7/requirements.txt
numpy==1.19.5
opencv-python==4.5.2.54
Pillow==8.2.0

